+++ @nav-frontend/ds-tokens/src/index.css
@@ -1,0 +1,2 @@
+@import "colors.css";
+@import "elevation.css";
