+++ @nav-frontend/popover/src/index.ts
@@ -1,0 +1,2 @@
+import Popover from "./popover";
+export default Popover;
