+++ package.json
@@ -61,6 +61,7 @@
     "url": "https://github.com/navikt/nav-frontend-moduler.git"
   },
   "dependencies": {
+    "@popperjs/core": "^2.5.4",
     "babel-plugin-react-docgen": "^1.8.1",
     "chalk": "^1.1.3",
     "css-loader": "^4.3.0",
@@ -82,6 +83,7 @@
     "pretty-quick": "^3.1.0",
     "react-docgen": "^5.3.0",
     "react-docgen-exports-handler": "^1.0.0",
+    "react-popper": "^2.2.4",
     "react-svg-icon-generator-fork": "^1.0.2",
     "start-server-and-test": "^1.11.4",
     "style-loader": "^1.2.1",
